anomalies/requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
faiss-cpu==1.8.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
parso==0.8.3
patchcore @ git+https://github.com/amazon-science/patchcore-inspection.git
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
opencv-python-headless
wand
